giang

+++ DoAn2_ASP/DoAn2_ASP/Areas/Admin/Controllers/DonMuonSachController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using DoAn2_ASP.Models;
+using Microsoft.AspNetCore.Http;
 
 namespace DoAn2_ASP.Areas.Admin.Controllers
 {
@@ -20,38 +21,62 @@ namespace DoAn2_ASP.Areas.Admin.Controllers
         }
 
         // GET: Admin/DonMuonSach
+        public bool XacNhanRole()
+        {
+            var taikhoanID = HttpContext.Session.GetString("StMaSinhVien");
+            if (_context.TblTaiKhoan.Where(t => t.StMaSinhVien == taikhoanID).Where(b => b.InMaQuyenHan == 1).Count() > 0)
+            {
+                return true;
+            }
+            else return false;
+        }
         public async Task<IActionResult> Index()
         {
-            var qL_ThuVienContext = _context.TblDonMuonSach.Include(t => t.StMaSinhVienNavigation);
-            return View(await qL_ThuVienContext.ToListAsync());
+            if (XacNhanRole() == true)
+            {
+                var qL_ThuVienContext = _context.TblDonMuonSach.Include(t => t.StMaSinhVienNavigation).ThenInclude(t => t.StMaKhoaNavigation);
+                return View(await qL_ThuVienContext.ToListAsync());
+            }
+            return NotFound();
+            
         }
 
         // GET: Admin/DonMuonSach/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (XacNhanRole() == true)
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var tblDonMuonSach = await _context.TblDonMuonSach
-                .Include(t => t.StMaSinhVienNavigation)
-                .FirstOrDefaultAsync(m => m.StMaDonMuon == id);
-            if (tblDonMuonSach == null)
-            {
-                return NotFound();
+                var tblDonMuonSach = await _context.TblDonMuonSach
+                    .Include(t => t.StMaSinhVienNavigation)
+                    .FirstOrDefaultAsync(m => m.StMaDonMuon == id);
+                if (tblDonMuonSach == null)
+                {
+                    return NotFound();
+                }
+                var chitiet = _context.TblChiTietDonMuon.AsNoTracking().Include(t => t.StMaSachNavigation).Where(m => m.StMaDonMuon == id).ToList();
+                ViewBag.chitiet = chitiet;
+                ViewData["StMaSinhVien"] = new SelectList(_context.TblSinhVien, "StMaSinhVien", "StMaSinhVien", tblDonMuonSach.StMaSinhVien);
+                return View(tblDonMuonSach);
             }
-            var chitiet = _context.TblChiTietDonMuon.AsNoTracking().Include(t => t.StMaSachNavigation).Where(m => m.StMaDonMuon == id).ToList();
-            ViewBag.chitiet = chitiet;
-            ViewData["StMaSinhVien"] = new SelectList(_context.TblSinhVien, "StMaSinhVien", "StMaSinhVien", tblDonMuonSach.StMaSinhVien);
-            return View(tblDonMuonSach);
+            return NotFound();
+            
         }
 
         // GET: Admin/DonMuonSach/Create
         public IActionResult Create()
         {
-            ViewData["StMaSinhVien"] = new SelectList(_context.TblSinhVien, "StMaSinhVien", "StMaSinhVien");
-            return View();
+            if (XacNhanRole() == true)
+            {
+                ViewData["StMaSinhVien"] = new SelectList(_context.TblSinhVien, "StMaSinhVien", "StMaSinhVien");
+                return View();
+            }
+            return NotFound();
+            
         }
 
         // POST: Admin/DonMuonSach/Create
@@ -61,31 +86,42 @@ namespace DoAn2_ASP.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("StMaDonMuon,StMaSinhVien,StMaKhoa,DaNgayMuon,DaNgayTra,BiTrangThai")] TblDonMuonSach tblDonMuonSach)
         {
-            if (ModelState.IsValid)
+            if (XacNhanRole() == true)
             {
-                _context.Add(tblDonMuonSach);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(tblDonMuonSach);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                ViewData["StMaSinhVien"] = new SelectList(_context.TblSinhVien, "StMaSinhVien", "StMaSinhVien", tblDonMuonSach.StMaSinhVien);
+                return View(tblDonMuonSach);
             }
-            ViewData["StMaSinhVien"] = new SelectList(_context.TblSinhVien, "StMaSinhVien", "StMaSinhVien", tblDonMuonSach.StMaSinhVien);
-            return View(tblDonMuonSach);
+            return NotFound();
+
+            
         }
 
         // GET: Admin/DonMuonSach/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (XacNhanRole() == true)
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var tblDonMuonSach = await _context.TblDonMuonSach.FindAsync(id);
-            if (tblDonMuonSach == null)
-            {
-                return NotFound();
+                var tblDonMuonSach = await _context.TblDonMuonSach.FindAsync(id);
+                if (tblDonMuonSach == null)
+                {
+                    return NotFound();
+                }
+                ViewData["StMaSinhVien"] = new SelectList(_context.TblSinhVien, "StMaSinhVien", "StMaSinhVien", tblDonMuonSach.StMaSinhVien);
+                return View(tblDonMuonSach);
             }
-            ViewData["StMaSinhVien"] = new SelectList(_context.TblSinhVien, "StMaSinhVien", "StMaSinhVien", tblDonMuonSach.StMaSinhVien);
-            return View(tblDonMuonSach);
+            return NotFound();
+            
         }
 
         // POST: Admin/DonMuonSach/Edit/5
@@ -95,52 +131,62 @@ namespace DoAn2_ASP.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("StMaDonMuon,StMaSinhVien,StMaKhoa,DaNgayMuon,DaNgayTra,BiTrangThai")] TblDonMuonSach tblDonMuonSach)
         {
-            if (id != tblDonMuonSach.StMaDonMuon)
+            if (XacNhanRole() == true)
             {
-                return NotFound();
-            }
+                if (id != tblDonMuonSach.StMaDonMuon)
+                {
+                    return NotFound();
+                }
 
-            if (ModelState.IsValid)
-            {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(tblDonMuonSach);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!TblDonMuonSachExists(tblDonMuonSach.StMaDonMuon))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(tblDonMuonSach);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!TblDonMuonSachExists(tblDonMuonSach.StMaDonMuon))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+                ViewData["StMaSinhVien"] = new SelectList(_context.TblSinhVien, "StMaSinhVien", "StMaSinhVien", tblDonMuonSach.StMaSinhVien);
+                return View(tblDonMuonSach);
             }
-            ViewData["StMaSinhVien"] = new SelectList(_context.TblSinhVien, "StMaSinhVien", "StMaSinhVien", tblDonMuonSach.StMaSinhVien);
-            return View(tblDonMuonSach);
+            return NotFound();
+            
         }
 
         // GET: Admin/DonMuonSach/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (XacNhanRole() == true)
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var tblDonMuonSach = await _context.TblDonMuonSach
-                .Include(t => t.StMaSinhVienNavigation)
-                .FirstOrDefaultAsync(m => m.StMaDonMuon == id);
-            if (tblDonMuonSach == null)
-            {
-                return NotFound();
-            }
+                var tblDonMuonSach = await _context.TblDonMuonSach
+                    .Include(t => t.StMaSinhVienNavigation)
+                    .FirstOrDefaultAsync(m => m.StMaDonMuon == id);
+                if (tblDonMuonSach == null)
+                {
+                    return NotFound();
+                }
 
-            return View(tblDonMuonSach);
+                return View(tblDonMuonSach);
+            }
+            return NotFound();
+            
         }
 
         // POST: Admin/DonMuonSach/Delete/5
@@ -148,10 +194,15 @@ namespace DoAn2_ASP.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var tblDonMuonSach = await _context.TblDonMuonSach.FindAsync(id);
-            _context.TblDonMuonSach.Remove(tblDonMuonSach);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (XacNhanRole() == true)
+            {
+                var tblDonMuonSach = await _context.TblDonMuonSach.FindAsync(id);
+                _context.TblDonMuonSach.Remove(tblDonMuonSach);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return NotFound();
+            
         }
 
         private bool TblDonMuonSachExists(int id)
